don't add lisp/other to load-path
Adding file .nosearch in that directory makes it less likely that it
is accidentally added to the load-path.  Users can still add it there
manually.  Some of the files in this directory are also available
separately and it can be really confusing for users when they install
a package explicitly but an older version bundled with some other
package is loaded instead.